+++ mokophone/callsdb.h
@@ -12,7 +12,7 @@ typedef enum
 
 struct _CallEntry {
     /* id chiamata */
-    gint64 id;
+    int id;
 
     /* direzione chiamata */
     CallDirection direction;
@@ -45,9 +45,9 @@ typedef void (*CallEntryFunc)(CallEntry*, gpointer);
 
 void callsdb_foreach_call(CallEntryFunc func, gpointer data);
 
-void callsdb_set_call_new(gint64 id, gboolean is_new);
+void callsdb_set_call_new(int id, gboolean is_new);
 
-gboolean callsdb_delete_call(gint64 id);
+gboolean callsdb_delete_call(int id);
 
 gboolean callsdb_truncate(void);
